月別配当のリファクタ
