+++ src/main/java/click/divichart/service/DividendService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.time.Year;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.StringJoiner;
@@ -25,6 +26,22 @@ public class DividendService {
     @Autowired
     public DividendService(DividendHistoryRepository dividendHistoryRepository) {
         this.repository = dividendHistoryRepository;
+    }
+
+    /**
+     * 指定された文字列の年を整数として取得します。
+     *
+     * <p>引数が空文字の場合は、現在の年を返します。それ以外の場合は、文字列をYearとして解析し、
+     * その年を整数値として返します。</p>
+     *
+     * @param targetYear 年を表す文字列。空文字の場合は現在の年を返す。
+     * @return 指定された年の整数値。targetYearが空の場合は現在の年。
+     */
+    public int getTargetYear(String targetYear) {
+        if (targetYear.isEmpty()) {
+            return Year.now().getValue();
+        }
+        return Year.parse(targetYear).getValue();
     }
 
     /**
