+++ src/main/java/click/divichart/service/YearlyCumulativeDividendService.java
@@ -4,7 +4,6 @@ import click.divichart.repository.DividendHistoryRepository;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
-import java.time.Year;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -14,22 +13,6 @@ public class YearlyCumulativeDividendService extends DividendService {
 
     public YearlyCumulativeDividendService(DividendHistoryRepository dividendHistoryRepository) {
         super(dividendHistoryRepository);
-    }
-
-    /**
-     * 指定された文字列の年を整数として取得します。
-     *
-     * <p>引数が空文字の場合は、現在の年を返します。それ以外の場合は、文字列をYearとして解析し、
-     * その年を整数値として返します。</p>
-     *
-     * @param targetYear 年を表す文字列。空文字の場合は現在の年を返す。
-     * @return 指定された年の整数値。targetYearが空の場合は現在の年。
-     */
-    public int getTargetYear(String targetYear) {
-        if (targetYear.isEmpty()) {
-            return Year.now().getValue();
-        }
-        return Year.parse(targetYear).getValue();
     }
 
     /**
